Took out api call from WO file and moved to API folder
api url is now getting pulled from the .env file
